fix: 여행 정보 검색 옵션 컴포넌트 분리
